Add files via upload
An XIA project using LIME where the Random Forest and Linear Regression models are discussed.
The databases have been provided. The text files of the code have been provided too.
It is an HR management project which explains how to train, compare, test for different objects and explain the observations.